Use quad_random instead of random
Otherwise wasm builds don't work

Ok, this is an interesting one. Apparently Macroquad has
a custom WASM build process[1], and the moment another WASM compatible
library is added the the project, it no longer works in browser.
rand crate, or rather its dependency getrandom, is one of those.

There is a way to make it potentially work together[2], but I decided
not going into this madness, I feel it's a huge time sink.

Using quad_rand is not without its problems, it still needs a source
of randomness to intialize the RNG; the workaround that I found is
using time from the launch of the app till the player chooses side.

Then, the shuffle method is Fisher-Yates algorithm that I got
from wikipedia[3]. Admittedly, it took more than one attempt to
weed out all the off-by-one errors that I made.

Overall I feel like this whole situation is a big turn-off for me
from using macroquad for anything that needs to run in a web-browser.
Being locked out from using many dependencies is quite a limitation.
As I mentioned above, it's potentially solvable, but this not something
you want to spend your time while developing a game.

[1] https://github.com/not-fl3/macroquad/issues/212
[2] https://github.com/not-fl3/macroquad/issues/212#issuecomment-835276147
[3] https://en.wikipedia.org/wiki/Fisher%E2%80%93Yates_shuffle